// assign14_15loop/src/authorization/permission.enum.ts
export const  enum PermissionKey {
  ViewUsers = 'ViewUsers',
  AddUsers = 'AddUsers',
  UpdateUsers = 'UpdateUsers',
  DeleteUsers = 'DeleteUsers',


  ViewCustomers = 'ViewCustomers',
  CreateCustomers = 'AddCustomers',
  UpdateCustomers = 'UpdateCustomers',
  DeleteCustomers = 'DeleteCustomers',

  SignUp = 'SignUp',
  LogIn = 'LogIn'


}